cli: Don't echo the password

+++ wmal/ui/cli.py
@@ -497,8 +497,9 @@ class wmal_accounts(AccountManager):
                 available_libs = ', '.join(sorted(utils.available_libs.iterkeys()))
                 
                 print "--- Add account ---"
+                import getpass
                 username = raw_input('Enter username: ')
-                password = raw_input('Enter password: ')
+                password = getpass.getpass('Enter password (no echo): ')
                 api = raw_input('Enter API (%s): ' % available_libs)
                 
                 try:
@@ -510,7 +511,10 @@ class wmal_accounts(AccountManager):
                 print "--- Delete account ---"
                 num = raw_input('Account number to delete: ')
                 num = int(num)
-                self.delete_account(num)
+                confirm = raw_input("Are you sure you want to delete account %d (%s)? [y/N] " % (num, self.get_account(num)['username']))
+                if confirm.lower() == 'y':
+                    self.delete_account(num)
+                    print 'Account %d deleted.' % num
             elif key.lower() == 'q':
                 sys.exit(0)
             else:
